refactor(release): split affected-native proof CLI
